# Build global ordinals terms bucket from matching ordinals