@pakein 2018-04-26
Get globalOrd by `globalOrd = bucketOrds.get(bucketOrd)`
 if globalOrd == 0,  it may be 
 bucket exists and globalOrd=0 
 or
 bucket not exists
 when globalOrd =0 match , it seems the result my be duplicate ?