@jimczi 2018-04-26
We use `bucketOrds#get` on a bucket ord that is guaranteed to exist since we only iterate the existing slots in `LongHash` and the ids in the hash are dense. The returned long is undefined if the slot is empty so `0` does not indicate a non-existing bucket but we should be safe here since we always use the filled slot.