@mattweber 2018-05-02
Thank you @jimczi and @pakein!  

I came across this issue when investigating a performance issue for one of my customers aggregations and applied this patch to a forked terms aggregation in a plugin for ES `6.2.1`.  For our use-case which is 4-levels of terms aggregations against ~3M docs and the 4th level being a very high cardinality field we went from ~350s down to ~20s!!

Any chance of getting this in before `6.4.0` as it has the potential for such massive performance improvements with no side-affects?