@jimczi 2018-05-03
> Any chance of getting this in before 6.4.0 as it has the potential for such massive performance improvements with no side-affects?

It was too late for 6.3.0 and it's not a critical bug (though this is arguable ;) ) hence 6.4.0.

> For our use-case which is 4-levels of terms aggregations against ~3M docs and the 4th level being a very high cardinality field we went from ~350s down to ~20s!!

Nice numbers ! Though 20s for 3M docs seems quite high even with 4 levels. How many terms do you retrieve per level ?