@mattweber 2018-05-03
@jimczi Thank you, `6.3.1` is perfectly fine and I understand it is not guaranteed to happen.

For our use-case, the first 3-levels are 100, and the 4th is unbounded (essentially all doc ids for the bucket).  Its this 4th level id collection that is the issue and I am working on trying to get them to remove this "requirement".