@fredgalvao 2018-05-26
Do {the tag swap + 6.3 commit} mean anything in regards to the possibility of having this in 6.3.0? Did anything change?
This is such a huge win and amazing optimization!