@jimczi 2018-05-28
@fredgalvao yes this change will be in 6.3.0